MidCryptosy/requirements.txt
cors                              1.0.1
Django                            4.0
django-cors-headers               3.12.0
djangorestframework               3.13.1
h5py                              3.7.0
joblib                            1.1.0
keras                             2.9.0
Keras-Preprocessing               1.1.2
kiwisolver                        1.3.2
langdetect                        1.0.9
libclang                          14.0.1
Markdown                          3.3.7
matplotlib                        3.5.0
nltk                              3.7
numpy                             1.21.4
oauthlib                          3.2.0
opencv-python                     4.5.4.58
opt-einsum                        3.3.0
packaging                         21.3
pandas                            1.4.2
patsy                             0.5.2
Pillow                            8.4.0
pip                               22.1.2

protobuf                          3.19.4
psycopg                           3.0.11
psycopg2                          2.9.3
pyasn1                            0.4.8
pyasn1-modules                    0.2.8
pycodestyle                       2.8.0
pycoingecko                       2.2.0
pycountry                         22.3.5
pycparser                         2.21
pyparsing                         3.0.6
PySocks                           1.7.1
python-dateutil                   2.8.2
pytz                              2022.1
regex                             2022.4.24
requests                          2.27.1
requests-file                     1.5.1
requests-oauthlib                 1.3.1
rsa                               4.8
schedule                          1.1.0
scikit-learn                      1.1.1
scipy                             1.8.1
setuptools                        58.1.0
setuptools-scm                    6.3.2
six                               1.16.0
sklearn                           0.0
sqlparse                          0.4.2
statsmodels                       0.13.2
tenacity                          8.0.1
tensorboard                       2.9.1
tensorboard-data-server           0.6.1
tensorboard-plugin-wit            1.8.1
tensorflow                        2.9.1
tensorflow-estimator              2.9.0
tensorflow-io-gcs-filesystem      0.26.0
termcolor                         1.1.0
textblob                          0.17.1
threadpoolctl                     3.1.0
tldextract                        3.3.0
toml                              0.10.2
tomli                             1.2.2
tqdm                              4.64.0
tweepy                            3.10.0
Werkzeug                          2.1.2
wheel                             0.37.1
wrapt                             1.11.2
zope.event                        4.5.0
zope.interface                    5.4.0